requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
celery==5.3.4
redis==5.0.1
numpy>=1.24.0
reportlab>=4.0.0
google-genai>=1.0.0
